fix: Move count_args func from builtin to utils directory.
